requirements.txt
appdirs==1.4.4
asgiref==3.3.4
astroid==2.5.2
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
coloredlogs==15.0
distlib==0.3.1
Django==3.2.4
et-xmlfile==1.1.0
filelock==3.0.12
humanfriendly==9.1
idna==2.10
iqoptionapi==5.2
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
MouseInfo==0.1.3
openpyxl==3.0.7
Pillow==8.2.0
PyAutoGUI==0.9.52
PyGetWindow==0.0.9
pylint==2.7.2
PyMsgBox==1.0.9
pyperclip==1.8.2
pyreadline==2.1
PyRect==0.1.4
PyScreeze==0.1.27
PyTweening==1.0.3
pytz==2021.1
pywin32==301
requests==2.25.1
selenium==3.141.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.5
virtualenv==20.4.4
websocket-client==0.56.0
wrapt==1.12.1
xlrd==2.0.1

PyQt5~=5.15.4